requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-framework-keycloak==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
drf-yasg==1.13.0
ecdsa==0.13
future==0.17.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
python-jose==3.0.0
python-keycloak==0.14.4
pytz==2018.9
requests==2.21.0
rsa==4.0
ruamel.yaml==0.15.87
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
